ATG: move OverlayManager to root
