+++ ATG/app/src/main/java/com/kuhakupixel/atg/ui/menu/AddressTableMenu.kt
@@ -15,7 +15,7 @@ import com.kuhakupixel.atg.backend.ACE.MatchInfo
 import com.kuhakupixel.atg.backend.ACE.NumType
 import com.kuhakupixel.atg.backend.ACEBaseClient
 import com.kuhakupixel.atg.ui.GlobalConf
-import com.kuhakupixel.atg.ui.overlay.service.OverlayComposeUI.OverlayManager
+import com.kuhakupixel.atg.ui.overlay.OverlayManager
 import com.kuhakupixel.atg.ui.util.CreateTable
 
 class AddressInfo(val matchInfo: MatchInfo, val numType: NumType) {
