+++ ATG/app/src/main/java/com/kuhakupixel/atg/ui/overlay/service/OverlayHackingScreenController.kt
@@ -11,13 +11,11 @@ import androidx.compose.material.Text
 import androidx.compose.material3.Button
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
-import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.Modifier
 import com.kuhakupixel.atg.ui.HackingScreen
 import com.kuhakupixel.atg.ui.overlay.OverlayViewHolder
 import com.kuhakupixel.atg.ui.overlay.service.OverlayComposeUI.OverlayInfoDialog
-import com.kuhakupixel.atg.ui.overlay.service.OverlayComposeUI.OverlayManager
+import com.kuhakupixel.atg.ui.overlay.OverlayManager
 import com.kuhakupixel.atg.ui.theme.AtgTheme
 
 class OverlayHackingScreenController(
